fix: fix composer container restart and
... also add logging from containers on panic, tracing for the cargo
 code and common message bus connect fn to facilitate testing.

Composer containers were not restarting because the containers were
being created with two networks, the global test and a "fake" one.

Address this by using a network with the compose name with defaults
to the global one but allow using different networks by making sure
that we always clean up networks with our tests labels.