Add initial volume service and the rest...
... server and client bindings.
The volume service is comprised of the nexus and volume operations where
the nexus is the "dumb" part of the service which creates nexuses where
it's told and the volume is the slightly smarter version which attempts
to decide where to create replicas, place nexuses etc...

The volume will effectively be the brain of the composer testing and
act as the equivalent to moac as csi controller plugin for mozart.

todo: mozart will need a clean and simple way of starting all the
required control plane services. Also, in the test do we want to
"talk" to the services through rest or the message bus?