nexus: Reject writes if any child is a snapshot
Add a new read-only flag to Nexus and when adding a child to the
nexus, check if its name matches the scheme used for a snapshots and,
if so, set the read-only flag.

When a submitting a write IO, if the nexus is read-only, fail the IO.
This stops the IO being submitting to the children, failing and then
causing the children to be faulted.

The nvmf target is still read/write as SPDK does not appear to have a
straightforward way of changing what it advertises.

TODO: Address other add child codepaths
Issue: Kernel initiator considers the write to have succeeded even
though the nvme request cqe returns status 0x06 (internal error)